+++ app/src/src_nonfreenet/org/breezyweather/sources/imd/ImdResultConverter.kt
@@ -22,7 +22,7 @@ import breezyweather.domain.weather.model.Temperature
 import breezyweather.domain.weather.model.Wind
 import breezyweather.domain.weather.wrappers.DailyWrapper
 import breezyweather.domain.weather.wrappers.HourlyWrapper
-import org.breezyweather.common.extensions.getFormattedDate
+import org.breezyweather.common.extensions.getIsoFormattedDate
 import org.breezyweather.sources.imd.json.ImdWeatherResult
 import java.text.SimpleDateFormat
 import java.util.Calendar
@@ -37,7 +37,7 @@ internal fun getDailyForecast(
 ): List<DailyWrapper> {
     // Need to provide an empty daily list so that
     // CommonConverter.kt will compute the daily forecast items.
-    val dates = hourlyForecast.groupBy { it.date.getFormattedDate("yyyy-MM-dd", location) }.keys
+    val dates = hourlyForecast.groupBy { it.date.getIsoFormattedDate(location) }.keys
     val formatter = SimpleDateFormat("yyyy-MM-dd", Locale.ENGLISH)
     formatter.timeZone = TimeZone.getTimeZone(location.timeZone)
     val now = Calendar.getInstance(TimeZone.getTimeZone(location.timeZone))