+++ app/src/src_nonfreenet/org/breezyweather/sources/openweather/OpenWeatherResultConverter.kt
@@ -27,7 +27,7 @@ import breezyweather.domain.weather.wrappers.CurrentWrapper
 import breezyweather.domain.weather.wrappers.DailyWrapper
 import breezyweather.domain.weather.wrappers.HourlyWrapper
 import org.breezyweather.common.extensions.capitalize
-import org.breezyweather.common.extensions.getFormattedDate
+import org.breezyweather.common.extensions.getIsoFormattedDate
 import org.breezyweather.common.extensions.toDate
 import org.breezyweather.common.extensions.toDateNoHour
 import org.breezyweather.sources.openweather.json.OpenWeatherAirPollution
@@ -64,7 +64,7 @@ internal fun getDailyList(
     if (hourlyResult.isNullOrEmpty()) return emptyList()
     val dailyList = mutableListOf<DailyWrapper>()
     val hourlyListByDay = hourlyResult.groupBy {
-        it.dt!!.seconds.inWholeMilliseconds.toDate().getFormattedDate("yyyy-MM-dd", location)
+        it.dt!!.seconds.inWholeMilliseconds.toDate().getIsoFormattedDate(location)
     }
     for (i in 0 until hourlyListByDay.entries.size - 1) {
         val dayDate = hourlyListByDay.keys.toTypedArray()[i].toDateNoHour(location.javaTimeZone)