+++ app/src/src_nonfreenet/org/breezyweather/sources/smhi/SmhiResultConverter.kt
@@ -24,7 +24,7 @@ import breezyweather.domain.weather.model.WeatherCode
 import breezyweather.domain.weather.model.Wind
 import breezyweather.domain.weather.wrappers.DailyWrapper
 import breezyweather.domain.weather.wrappers.HourlyWrapper
-import org.breezyweather.common.extensions.getFormattedDate
+import org.breezyweather.common.extensions.getIsoFormattedDate
 import org.breezyweather.common.extensions.toDateNoHour
 import org.breezyweather.sources.smhi.json.SmhiTimeSeries
 import kotlin.math.roundToInt
@@ -35,7 +35,7 @@ internal fun getDailyForecast(
 ): List<DailyWrapper> {
     val dailyList = mutableListOf<DailyWrapper>()
     val hourlyListByDay = forecastResult.groupBy {
-        it.validTime.getFormattedDate("yyyy-MM-dd", location)
+        it.validTime.getIsoFormattedDate(location)
     }
     for (i in 0 until hourlyListByDay.entries.size - 1) {
         val dayDate = hourlyListByDay.keys.toTypedArray()[i].toDateNoHour(location.javaTimeZone)